move ReactionList component to src/pages/ReactionPage
